Move NoDataAnnotationsCustomization into the DataAnnotations namespace
