+++ Src/AutoFixture/DataAnnotations/NoDataAnnotationsCustomization.cs
@@ -1,18 +1,17 @@
 ﻿using System;
 using System.Linq;
-using AutoFixture.DataAnnotations;
 
-namespace AutoFixture
+namespace AutoFixture.DataAnnotations
 {
     /// <summary>
     /// A customization that removes support for generating values that are customized
-    /// based on the member's data annotations attributes. 
+    /// based on the member's data annotations attributes.
     /// </summary>
     /// <remarks>
     /// This removes the <see cref="RangeAttributeRelay"/>,
     /// <see cref="StringLengthAttributeRelay"/>, and
     /// <see cref="RegularExpressionAttributeRelay"/>. Support for the removal of data
-    /// annotations is provided for performance reasons, however a typical user of 
+    /// annotations is provided for performance reasons, however a typical user of
     /// AutoFixture would not need to worry about the performance.
     /// </remarks>
     public class NoDataAnnotationsCustomization : ICustomization
